requirements.txt
load-dotenv==0.1.0
telethon==1.40.0

